msys2-runtime: Add new patch for path_conv
Use whichever mount gives the longest result.  This change
is because argv[0] of MSYS2 programs was /bin/foo.exe when
it should be /usr/bin/foo.exe and this caused all sorts of
issues with relocation (as e.g. PREFIX is /usr and that is
expected to be found as the prefix of argv[0]).

For example, from cmd.exe:

"C:\msys64\usr\bin\g++.exe -print-prog-name=cc1plus"

.. gave:

"cc1plus"

.. before this patch, when it should have given:

"/usr/lib/gcc/x86_64-pc-msys/4.8.2/cc1plus.exe"