Fix building the MSYS2 installer
The available pacman executable does not know the `-rootrebase` option,
therefore we need to use the `-root` option.

This developer suggests that the `-rootrebase` option was introduced to
ask pacman to rebase the .dll files after installing everything into a
new root directory. However, we only ever plan to develop Git for
Windows on 64-bit setups, therefore rebasing is substantially less
crucial for us and we can totally live without it.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>