tcl: New package
Uses a patch from Cygwin and is at the same
version. We won't be providing tk though as
that requires the X Windows System - unless
someone reverts the change that removed GDI
as a rendering back-end for Cygwin's tk.

This was added in support of expect which
was in turn added in support of dejagnu so
that libffi's testsuite can be run.