Include the 'binutils' package into base-devel
The recommended way to start building MSys2 packages is to call

	pacman -S base-devel

once and then run `makepkg -s` in the respective subdirectories of the
MSYS2-packages project.

Many of these packages require the `strip` utility which is provided by
the `binutils` package, and as a matter of fact, `makepkg` refuses to
start if it does not find the `strip` executable.

So let's just include the `binutils` package in the `base-devel` groups so
that it gets installed in the recommended workflow.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>