vim: accept CR/LF in input (such as ~/.vimrc)
When sharing Windows' home directory, we have to expect that the
`.vimrc` file can have DOS line endings.

Let's be lenient about this, as well as all other input we get, and just
handle CR/LF as well as LF line endings.

We still behave as before, otherwise.

This fixes https://github.com/git-for-windows/git/issues/272 and
https://github.com/git-for-windows/git/issues/364

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>