cyrus-sasl: Remove annoying debug message
Every time an svn call is made, the message "DIGEST MD5 common mech free"
appears in the application event log viewer. Other communities (See
https://www.mail-archive.com/debian-bugs-dist@lists.debian.org/msg1372451.html)
are supressing the message with a logcheck snippet. Since we do not have
such a luxury in the msys2 environment, lets just remove the debug message
for good.

This fixes https://github.com/git-for-windows/git/issues/616

Signed-off-by: 마누엘 <nalla@hamal.uberspace.de>