msys2-runtime: fix build
Upstram MSYS2 finds it funny to move from a perfectly fine Git branch to
shipping gazillions of patches (that are hard to fix in case they no
longer apply).

Git for Windows' maintainer (yours truly, pleasure to meet you) disagrees
and keeps things in a Git branch that is kept up-to-date via merging
rebases.

That means that we diverge substantially from upstream, but also that
things are easier for us.

This patch reverts the part from upstream's process that was merged by
mistake.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>