msys2-runtime: bring it closer to upstream MSYS2 again
Upstream decided to go to the patch-based approach again, applying
directly on top of Cygwin's releases. The msys2-master branch still
reflects the exact patches, so it is *also* Git-based, but you have to
look a little harder.

Let's go there, too, but already rebase on top of Cygwin 2.6.1.

As always, this rebase was performed using the Git garden shears, as a
"merging rebase", i.e. both a merge (retaining the history) and a rebase
(replaying all the patches on top of the merge that essentially wiped out
all changes for that purpose).

The resulting patches were exported using this command-line:

	git format-patch -o ../.. \
		cygwin-2_6_1-release..  ^HEAD^{/Start.the.merging-rebase}

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>