msys2-runtime: add helper script to update the patch set
It is a bit tedious, in particular with the recent switch from a Git
branch in https://github.com/Alexpux/Cygwin to a patch set, to update
the msys2-runtime.

Let's simplify things again by scripting the process that turns real Git
commits into that patch set.

It is still a bit cumbersome to keep track of the MSYS2 runtime patches
since the msys2-master branch is always rebased, with no record of prior
iterations, of course.

This script has been contributed in almost the same shape as part of
https://github.com/Alexpux/MSYS2-packages/pull/786, with one modification:
the current version handles Git for Windows' merging rebases [*1*]
gracefully.

Footnote *1*: "Merging rebases" are Git for Windows' strategy to rebase
*and* retain the previous commit history. For details, please see
https://github.com/git-for-windows/build-extra/blob/4cacaabc1/shears.sh#L12-L18

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>