msys2-runtime: fix some problems with the POSIX->Windows munging
There was a subtle buffer overrun in the convert() mechanism: after a
sub_convert() completed, we may very well be at the end of the
conversion, i.e. the "src" pointer may already point at the trailing
NUL. Advancing that pointer any further shifts the pointer out of
bounds, with the predictable result of copying bogus characters into the
output buffer.

Another very subtle buffer overrun happens when the the buffer passed to
the find_path_start_and_type() function ends in a colon: in that case,
the code advances one byte to see whether there is a slash, another
colon, or a period, or if there is no equal sign in th rest of the
buffer, but since there is only a NUL byte, the code enters the next
loop iteration, advancing by one byte *again*, and consequently failing
the `it2 != end` condition.

While these bugs were not really reproducible in this developer's hands
(the only way to reproduce a problem was in a very specific setup: a
Visual Studio Team Services build agent running Git for Windows' full
test suite in parallel, triggered as build job, not manually on the
Virtual Machine), they still need to be fixed.

This change corresponds to commit 3cf1b9c3ac (fixup! Add functionality
for converting UNIX paths in arguments and environment variables to
Windows form for native Win32 applications., 2017-01-06) in
https://github.com/git-for-windows/msys2-runtime (which keeps all
previous history even when rebasing to the newest Cygwin/MSYS2 runtime).

Unfortunately, the diff is quite big due to the way MSYS2 runtime's
patches are handled now: even one change in one patch changes all the
subsequent .patch files.

To make it easier to validate this change, the corresponding Git history
has been tagged as pr-786-step-1 in https://github.com/dscho/msys2-runtime.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>