msys2-runtime: backport FAST_CWD fix for upcoming Windows versions
A *very* recent Windows build adds more code to the preamble of
RtlGetCurrentDirectory_U() so that the previous heuristic failed to find
the call to the locking routine.

This only affects the 64-bit version of ntdll, where the 0xe8 byte is now
found at offset 40, not the 32-bit version. However, let's just double the
area we search for said byte for good measure.

This patch has already been accepted into Cygwin's source code.

Unfortunately, due to the way MSYS2 runtime's changes are handled these
days, even the addition of one patch changes the *entire* patch set, as
the "Subject:" lines have to increase the total number of patches.

To make validation of this patch easier, the corresponding Git history has
been tagged as pr-786-step-2 in https://github.com/dscho/msys2-runtime.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>