msys2-runtime: fix "Bad address" problem in the path conversion
There was an infinite recursion when a POSIX path list contained the
substring `://` that is not really part of a URL: `ppl_convert()` would
mis-identify this as a URL, pass it to `subp_convert()` which would then
call a function that would then mis-identify the component as a POSIX
path list and the game would start anew.

This infinite recursion was reported as "Bad address" to the user,
causing no end of confusion.

The first two patches added in this commit address this infinite
recursion, and the URL mis-identification, respectively.

While in the area already, the third patch reverts a Git for
Windows-specific patch that is no longer needed to let Git for Windows'
test suite pass: just like the MSYS2 runtime, we now append a trailing
slash when converting the root directory `/` to a Windows path.

This fixes https://github.com/git-for-windows/git/issues/1033

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>