msys2-runtime: adjust patch set
The patches on top of the Cygwin runtime were regenerated from Git for
Windows' fork at https://github.com/git-for-windows/git. As that
repository uses the merging-rebase strategy, the commits are stable,
i.e. the commit IDs recorded in the .patch files will not go away with
the next rebase, but will stay available eternally.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>