Remove last remnant of the msys2-installer
A long, long, *long* time ago we thought that the MSYS2 project may be
interested in collaborating on a better installer, one that we could
adjust easily to install the Git for Windows SDK.

That collaboration turned out a little too hard to be feasible, for both
sides, therefore Git for Windows simply resurrected the SDK installer
from the Git for Windows 1.x days and adjusted it minimally to install
the MSYS2-based setup.

Therefore we really need nothing from the msys2-installer/ directory any
longer, and as MSYS2 itself decided to move that directory into an own
repository, it can leave our fork, too.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>