msys2-runtime: use a more resilient method to update the patches
Whenever a new patch is added, or one is removed, *all* patches change
because of the [PATCH M/N] prefix of the Subject: line.

That is unnecessary because we actually do not care *how many* there
are, we can easily just look at the file list to find out.

So let's just replace the "/N" part by a literal letter N.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>