msys2-runtime: bump to 2.8.0pre1
Technically, Cygwin did not release 2.8.0 yet. But they published a test
release: https://cygwin.com/ml/cygwin-announce/2017-03/msg00026.html,
and their plan is to release some time this week.

Normally, we would wait, but Git v2.12.1 was released yesterday, and we
do not want to delay the release of Git for Windows v2.12.1, yet we also
do not want to stay on the MSYS2 runtime 2.7.0, as it seems that there
may be adverse interactions with a recent Windows update that appear to
have been addressed by fixes in 2.8.0 (for full details, see
https://cygwin.com/ml/cygwin/2017-03/msg00113.html).

Therefore, let's just cut a "pre" version for now (the version string
was crafted specifically so that Pacman will understand that 2.8.0 is a
newer version).

The only merge conflict that arose during the rebase was caused by the
fact that Cygwin newly exports the timingsafe_bcmp, timingsafe_memcmp,
and the dladdr symbols (and we still add the kill_process_tree symbol to
the fray).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>