gnupg: introduce GPG_ASKPASS support
It is all fine and dandy when users GPG-sign files from the terminal.
But when they do not, things get more complicated.

The usual thing is to install a pinentry helper targeting your graphical
desktop and then launching the GPG agent somehow and connet it all
together.

Except that MSYS2 does not come with a gpg-agent.

So let's just imitate openssh, which interprets the environment variable
SSH_ASKPASS (if set) as the path to a helper that can be called upon to
ask the user to provide a passphrase.

Unsurprisingly, the new environment variable is called GPG_ASKPASS (and
we do not repeat Git's mistake of co-opting SSH_ASKPASS, if set).

This allows, say, Git for Windows' Git GUI to commit with GPG signatures
even if the passphrase is not empty.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>