msys2-runtime: reintroduce `this != NULL` checks
While GNU C++ thinks of these as serious programming errors, the truth
is: those checks are not cosmetic. The reason is that we sometimes
execute methods via pointers that *sometimes* happen to be NULL.

So those checks are essential and we have no business removing them.

To make things compile again, we specifically ask GNU C++ very politely
and in the nicest manner to please shut up already about those `this !=
NULL` checks.

This fixes https://github.com/git-for-windows/git/issues/1316

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>