curl: prepare MSYS2 patch for v7.58.0
There were conflicts, and it is definitely time to make this much easier
to support by starting to use Git effectively.

So here we go, a new patch, exported from the freshly-pushed
https://github.com/dscho/curl/tree/msys2-curl-7.58.0

This branch is based on the tar import of cURL v7.58.0, and has
cherry-picked (with merge conflicts resolved) the tip of
https://github.com/dscho/curl/tree/msys2-curl-7.57.0 (which was the
previous version of the patch, now deleted).

It is probably not a wise idea to encode the cURL version in the file
name of the patch, but at least it was not introduced by yours truly.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>