heimdal: try harder to work around the absence of `ln -s`
Heimdal uses a really dirty trick to pretend that there is a hcrypto/
directory with includes: it adds a symbolic link of the current
directory into a subdirectory with the name "hcrypto".

On MSYS2, we do not yet have a reliable way to emulate `ln -s`, so we
try `ln` instead (which fails).

The previous patch only worked around this half-way, by skipping that
`ln` call.

With this updated patch, the correct `-I..` flag is also passed into the
compiler so that it finds the respective directory in the include path.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>