perl: drop unnecessary dependencies
Over the course of Git for Windows 2.*, the installer has grown from
~28MB to ~38MB. This is a bit sad because we really want it to be as
small a download as possible (since it is served from GitHub Releases,
backed by AWS, downloads can be really slow e.g. in South East Asia).

In Git for Windows, we will want to have as lean and mean of a Perl as
possible anyway. And since we don't need any database functionality...
out with it!

We will try to remove even more stuff when bundling the installer, as we
certainly won't need things like the MakeMaker (end user installations
of Git for Windows do not come with a C compiler...).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>