openssl: upgrade to v1.1.1a
Only the ca-dir patch applied cleanly. The MSYS2 one had to be re-done,
as well as the enginesdir one (where we now completely side-step
./configure).

This will require *tons* of packages to be rebuilt, and during the
transition time the binaries will be broken because of the different
name of the .dll, unless one copies back msys-crypto-1.0.0.dll and
msys-ssl-1.0.0.dll (this is how Git for Windows handled that fragile
upgrade phase).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>