DO NOT MERGE - This does NOT yet work. This is an attempt to build MSYS @ run-time off of the newlib 3.0 version.  Unfortunately, after installing, it kept complaining about being unable to find libraries and executables.  I'm missing something with these patches..  THe version of newlib I was trying to use was  3.0.4.  In addition, I also got a warning about a  missing "_wait" function with newlib.
/home/jpmugaas/msys/msys2-runtime/src/msys2-runtime/newlib/libc/reent/execr.c: In function '_wait_r':
/home/jpmugaas/msys/msys2-runtime/src/msys2-runtime/newlib/libc/reent/execr.c:120:14: warning: implicit declaration of function '_wait'; did you mean 'wait'? [-Wimplicit-function-declaration]
   if ((ret = _wait (status)) == -1 && errno != 0)
              ^~~~~
              wait

TBH, I am at my whits wned but I hope someone can complete this wwork and eventually get this merged because we probably should be more current with what is in CygWin.  They may have done things that we can benefit from.  In adidon, I am also trying to IFDEF the modifications made to the newlib code so that if the CygWin developers want to merge it, they can.    And it would help people wanting to follow what is happening between the versions of newlib.