rebaseall: include _all_ native Perl modules
There seem not to be any `.dll` files in the `site_perl/` directory
structure, probably for quite a while. However, there are quite a couple
in `core_perl/` (and the Subversion bindings are in `vendor_perl/`, if
installed).

Since we forgot to look for those when rebasing, it is not a big
surprise that some builds, notably builds using `autoconf`, would fail:

      0 [main] perl 5228 child_info_fork::abort: address space needed by 'Cwd.dll' (0x30000) is already occupied
Can't fork, trying again in 5 seconds at /usr/share/autoconf/Autom4te/General.pm line 307.

Let's just include those `.dll` files when rebasing.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>