pacman: kill all msys processes after a core upgrade
After a core upgrade instead of waiting endlessly and hoping that
the user closes all remaining msys processes ask the user if we should
kill them and then just do it.

This iterates over /proc/*/winpid and passes the list to taskkill
which kills all processes, except pacman itself which exits manually.

In case of --noconfirm things get killed automatically.