Add a base package
This is meant as a replacement for the base group and contains the
same packages. If a user has the base package installed (the installer
will do this in the future) then we can add new packages to the base
and users with existing installations will get those packages installed
automatically.

This follows what Arch did last year:
https://www.archlinux.org/news/base-group-replaced-by-mandatory-base-package-manual-intervention-required/

See #1976