openssl: Add a separate openssl-docs package
This contains the htmls docs/manpages and the level 3 man pages (API docs).
openssl is part of base and all these docs make up 7800 files, which is a third
of all files in base.

Since they are rarely needed and contain so many files I think it makes sense to split them out.