filesystem: Make msystem compatible with POSIX compliant shells. (#2009)
Behavior about the value of HOSTTYPE and MACHTYPE is not compatible
between various POSIX compliant shells. For example zsh doesn't set
HOSTTYPE at all and set MACHTYPE to that of HOSTTYPE in bash. I
checked source tree of bash and found that HOSTTYPE is set to the
result of `uname -m` and that MACHTYPE is set to the result of `uname
-m` plus "-pc-msys".So change code accordingly.

This fixes issue #2008.

Co-authored-by: Yasuhiro KIMURA <yasu@utahime.org>