filesystem: remove rege-info.sh
It calls install-info on all info files on the first MSYS2 start, but all
the info files are already installed in our current installer.

I'm not 100% sure why this was added, but I guess because of #1998 the installer
created in a chroot didn't install the info files properly and thus they were
missing on a fresh install. This is no longer the case now.

regen-info.sh is called in /etc/profile and part of the "backup" array in the PKGBUILD
so the call might not be removed if the user has changed /etc/profile. Since it is only
called in the first MSYS2 session this shouldn't lead to any problems.