CI: verify signatures of source files
With auto-key-retrieve enabled gpg called by makepkg will fetch keys
from validpgpkeys in the PKGBUILD and verify the result.
Uses the ubuntu keyserver because that one isn't as flaky.

This requires that all our packages get an up-to-date validpgpkeys
array. But we no longer have to trust the users to have checked the
signature before creating a PR.