automake: Don't depend on Python 2
This also removes the pgp checks for everything except 1.16 because
they don't contain the full key ID and can't be auto fetched by gnupg.
Since we have the checksums and they wont be updated in the future anyway
this shouldn't be a problem.

The tests didn't work anyway, so removing Python 2 doesn't change much.