Add the 'tmate' package
The `tmate` project forked the `tmux` code to allow for easy,
almost-instantaneous pair programming in the console.

Using this project, the GitHub Action
https://github.com/marketplace/actions/debugging-with-tmate allows
debugging build failures interactively.

Unfortunately, `tmate` does not offer any support for Windows.

MSYS2 to the rescue! We just added package definitions for `msgpack-c`
and `libssh`, which are the two previously missing dependencies of
`tmate`, and with this commit, we add the package definition for `tmate`
itself.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>