mingw-w64-cross-*: hardcode the commit hash in the version
Otherwise "git am" used in prepare() wil change HEAD and as a result pkgver().
Also when "git am" is used the hash in the version isn't in the upstream repo
which makes it hard to figure out where it was built from.

Just hardcode to the commit that was checkedd out initially.