libevent: fix DNS lookups in MSYS2
There is code in libevent to obtain the list of name server IPs the
Windows way, but it was disabled when building in MSYS mode.

To avoid conflicts between the Win32 and the MSYS headers, a hack is
needed where the `evutil_load_windows_system_library_()` function no
longer uses any Win32 types but plain C ones.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>