msys2-runtime: handle Ctrl+C with MINGW executables
The idea is that we handle Ctrl+C not by terminating the processes
(which would reflect `kill -KILL`, not `kill -INT`) but by imitating
what Windows does in a CMD window: we inject a remote thread that runs
the `CtrlRoutine` function. This lets the `atexit()` handlers and the
ConsoleCtrlEvent handlers do their job.

This commit brings the goodness from
https://github.com/msys2/msys2-runtime/pull/31 over into MSYS2-packages,
which cherry-picked and edited the patches from Git for Windows.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>