msys2-runtime: do not fall back to ExitProcess() on Ctrl+C
As long as the return value of `CtrlRoutine` indicates that it has been
handled, do not insist on the process to go away. This is in line with
`SIGINT` handlers on Linux/Unix/macOS that can prevent the process from
terminating on the process' side.

This brings over https://github.com/msys2/msys2-runtime/pull/50 into
MSYS2-packages.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>