pacman: try to improve makepkg-mingw defaults
* Print MINGW_ARCH info when passing --help/-h
* Change behaviour when MINGW_ARCH isn't set
  * Default to MSYSTEM if in a mingw environment
  * Default to mingw64 otherwise (but warn)
* Print some feedback on what is getting built right away.
  makepkg can take some time at the start and this makes
  things clear right away.

The goal of the MINGW_ARCH change is to make it more likely to do the right thing by
default and to not special case any environment.

This has the potential of breaking existing users if MINGW_ARCH isn't set
and they want exactly "mingw32/mingw64" to be built like before. This will now build
only mingw64 if called in a msys environment and either mingw32/64 if called
in those environments.

To restore the old behavior set MINGW_ARCH="mingw64 mingw32" first.