filesystem: fix nested `msys2_shell.cmd -full-path` calls
When calling `msys2_shell.cmd -full-path` from _within_ an MSYS2 Bash,
the `ORIGINAL_PATH` variable is set already. This will get converted to
a Windows-style path list when calling `cmd`, and not converted back
when starting the nested Bash. Hence, the `MSYS2_PATH_STYLE=inherit`
explicitly configured via `-full-path` will cause `ORIGINAL_PATH` to be
incorporated into the `PATH` variable with Windows-style paths.

This is incorrect.

As we already know that the `ORIGINAL_PATH` will contain a Windows-style
path list in the `.cmd` script (because it is executed via `cmd.exe`,
which is not an MSYS program), let's just invalidate it before
configuring the path style to `inherit`.

This fixes https://github.com/msys2/msys2-runtime/issues/43.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>