msys2-runtime: special case the SHELL and ORIGINAL_PATH env vars
In MSYS2 <-> MINGW transitions, we convert a couple of environment
variables between Unix and Windows styles, most notably `PATH`.

In the Git for Windows project, we realized that the environment
variables `SHELL` and `ORIGINAL_PATH` need the same treatment. These
patches have proved their worth for years, and it is time for them to be
integrated back into the actual MSYS2 runtime.

This should fix https://github.com/msys2/msys2-runtime/issues/43.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>