readline: avoid linking the static library to the ncurses DLL
Since
https://github.com/ThomasDickey/ncurses-snapshots/commit/ba50399b8b9cc69cb56d8a51754c4366b2ae3f7e#diff-e468d12af06630d1a5e6bca47f23638abdcf3eccfda06a629387934c5196bbeaL71-L76,
the ncurses headers no longer default to static linking, but to dynamic
linking.

This means that we now have to set the macro `NCURSES_STATIC` explicitly
when we plan on linking to said library statically.

Which we do, when linking the static version of the readline library
(which is used e.g. in Bash, where we definitely want to avoid linking
to any MSYS DLL except `msys-2.0.dll`).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>