subversion: drop a patch that breaks the test suite
This patch may have "fixed" the build at some stage when it failed with this symptom:

	[...]
	cd subversion/tests && /bin/sh "/usr/src/MSYS2-packages/subversion/src/subversion-1.14.2/libtool" --tag=CC --silent --mode=link gcc -shared  -march=x86-64 -mtune=generic -O2 -pipe   -march=x86-64 -mtune=generic -O2 -pipe  -pipe     -rpath /usr/lib -version-info 0  -o libsvn_test-1.la  svn_test_fs.lo svn_test_main.lo ../../subversion/libsvn_repos/libsvn_repos-1.la ../../subversion/libsvn_fs/libsvn_fs-1.la ../../subversion/libsvn_delta/libsvn_delta-1.la ../../subversion/libsvn_subr/libsvn_subr-1.la -laprutil-1 -lapr-1 -lintl
	libtool:   error: can't build x86_64-pc-msys shared library unless -no-undefined is specified                           make: *** [build-outputs.mk:794: subversion/tests/libsvn_test-1.la] Error 1

However, some tests expect to be located in the original location. For
example, `checksum-test.exe` needs to find `zlib.deflated`, but that is
not in `subversion/tests/`, it is in `subversion/tests/libsvn_subr/`.
Hence the `.exe` needs to live there, too.

Let's put in a hack that lets that library build _and_ the tests to
pass.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>