makepkg: don't use "-mno-*" in CFLAGS
In case a package has a configure check for say -mssse3 and then
enables it if the compiler supports it it would be passed to the compiler
before our CFLAGS, which currently has -mno-ssse3 (which seems to also disable
sse4 and above while at it). The code then assumes it's available but the compiler
fails.

While one could argue that we try to prevent ssse3+ being used in packages, there
are always exceptions like building sse4 code which dynamically gets executed.

So it's better we only enable machine options in CFLAGS and never disable them

Replace "-march=core2 -mno-ssse3" with the equivalent "-march=nocona -msahf".
Note that the GCC docs wrongly state that "nocona" doesn't include "CX16", while it does.

See https://github.com/msys2/MINGW-packages/pull/13651