msys2-runtime: update to v3.4.0
This reflects https://github.com/msys2/msys2-runtime/pull/120.

Note: As Cygwin changed their naming scheme, we had to adjust both the
source URL in the `PKGBUILD` file as well as the `update-patches.sh`
script to reflect that (both would have expected the Cygwin tag to be
named `cygwin-3_4_0-release` while it is now named `cygwin-3.4.0` which
is quite obviously a much better name, too).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>